requirements.txt
requests
pandas
pydantic
python-dotenv
pyyaml
rapidfuzz
pytest
streamlit
